# swagger.json
{"openapi":"3.0.0","info":{"title":"Library Management API","version":"1.0.0","description":"API documentation for the Library Management System"},"servers":[{"url":"http://localhost:3000","description":"Local server"}],"paths":{"/books":{"post":{"summary":"Add a new book","tags":["Books"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["title","author","isbn"],"properties":{"title":{"type":"string"},"author":{"type":"string"},"isbn":{"type":"string"},"quantity":{"type":"integer"},"shelf_location":{"type":"string"}}}}}},"responses":{"201":{"description":"Book created successfully"},"400":{"description":"Invalid request"}}},"get":{"summary":"Get all books","tags":["Books"],"responses":{"200":{"description":"List of books"}}}},"/books/search":{"get":{"summary":"Search for books by title, author, or ISBN","tags":["Books"],"parameters":[{"in":"query","name":"q","schema":{"type":"string"},"required":true,"description":"Search term (title, author, or ISBN)"}],"responses":{"200":{"description":"Matching books"},"429":{"description":"Too many requests (rate limited)"}}}},"/books/{id}":{"get":{"summary":"Get a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Book ID"}],"responses":{"200":{"description":"Book details"},"404":{"description":"Book not found"}}},"put":{"summary":"Update a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Book ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string"},"author":{"type":"string"},"isbn":{"type":"string"},"quantity":{"type":"integer"},"shelf_location":{"type":"string"}}}}}},"responses":{"200":{"description":"Book updated successfully"},"404":{"description":"Book not found"}}},"delete":{"summary":"Delete a book by ID","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Book ID"}],"responses":{"200":{"description":"Book deleted successfully"},"404":{"description":"Book not found"}}}},"/borrowed-books/borrowers/{id}/borrow":{"post":{"summary":"Borrow a book for a borrower","tags":["BorrowedBooks"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["book_id","due_date"],"properties":{"book_id":{"type":"integer"},"due_date":{"type":"string","format":"date"}}}}}},"responses":{"201":{"description":"Book borrowed successfully"},"400":{"description":"Invalid request"},"404":{"description":"Borrower or book not found"},"429":{"description":"Too many requests (rate limited)"}}}},"/borrowed-books/borrowers/{id}/return":{"put":{"summary":"Return a borrowed book","tags":["BorrowedBooks"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["book_id"],"properties":{"book_id":{"type":"integer"}}}}}},"responses":{"200":{"description":"Book returned successfully"},"400":{"description":"Invalid request"},"404":{"description":"Loan record not found"}}}},"/borrowed-books/borrowers/{id}/loans":{"get":{"summary":"List all books currently borrowed by a borrower","tags":["BorrowedBooks"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"responses":{"200":{"description":"List of borrowed books for the borrower"},"404":{"description":"Borrower not found"}}}},"/borrowed-books/loans/overdue":{"get":{"summary":"List all overdue borrowed books","tags":["BorrowedBooks"],"responses":{"200":{"description":"List of overdue loans"}}}},"/borrowers":{"get":{"summary":"Get all borrowers","tags":["Borrowers"],"responses":{"200":{"description":"List of borrowers"}}},"post":{"summary":"Add a new borrower","tags":["Borrowers"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","email"],"properties":{"name":{"type":"string"},"email":{"type":"string","format":"email"}}}}}},"responses":{"201":{"description":"Borrower created successfully"},"400":{"description":"Invalid request"}}}},"/borrowers/{id}":{"get":{"summary":"Get a borrower by ID","tags":["Borrowers"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"responses":{"200":{"description":"Borrower details"},"404":{"description":"Borrower not found"}}},"put":{"summary":"Update a borrower by ID","tags":["Borrowers"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string","format":"email"}}}}}},"responses":{"200":{"description":"Borrower updated successfully"},"404":{"description":"Borrower not found"}}},"delete":{"summary":"Delete a borrower by ID","tags":["Borrowers"],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Borrower ID"}],"responses":{"200":{"description":"Borrower deleted successfully"},"404":{"description":"Borrower not found"}}}}},"components":{},"tags":[{"name":"Books","description":"Book management"},{"name":"BorrowedBooks","description":"Borrowing process management"},{"name":"Borrowers","description":"Borrower management"}]}